pyproject.toml
[tool.poetry]
name = "mozanta-python-training-project"
version = "0.1.0"
description = ""
authors = ["sopal <sopal.soman@mozamta.com>"]
readme = "README.md"
packages = [{include = "mozanta-python-training-project"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.8"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
pillow = "^10.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
